+++ client/src/pages/main/Main.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import AppAppBar from '../../organisms/main/layout/AppAppBar';
 import AppFooter from '../../organisms/main/layout/AppFooter';
 import ProductHero from '../../organisms/main/Main/views/Hero/ProductHero';
@@ -9,47 +9,9 @@ import withRoot from '../../organisms/main/Main/withRoot';
 import sources from '../../organisms/main/Main/source/sources';
 import BetaDialog from '../../organisms/main/Main/views/Login/BetaDialog';
 
-import axios from '../../utils/axios';
-import HOST from '../../utils/config';
-import history from '../../history';
+import useLoginValue from '../../utils/lib/hooks/useLoginValue';
+import useDialog from '../../utils/lib/hooks/useDialog';
 
-const useLoginValue = () => {
-  const [isLogin, setisLogin] = useState(false);
-  const [repasswordOpen, setRepassword] = useState(false);
-
-  // logout function
-  const logout = () => {
-    setisLogin(false);
-    axios.get(`${HOST}/api/login/logout`)
-      .then(() => {
-        history.push('/');
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  useEffect(() => {
-    axios.get(`${HOST}/api/login/check`)
-      .then((res) => {
-        if (!res.data.error) {
-          if (res.data.state) {
-            // 임시 로그인되었습니다.
-            console.log('임시로그인 되었습니다.');
-            setRepassword(true);
-          }
-          setisLogin(true);
-        }
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  });
-
-  return {
-    isLogin, repasswordOpen, logout, setRepassword,
-  };
-};
 
 export default withRoot((props) => {
   // if located here, set the scroll to top of the page
@@ -58,10 +20,8 @@ export default withRoot((props) => {
     isLogin, repasswordOpen, logout, setRepassword,
   } = useLoginValue(location);
 
-  const [open, setOpen] = useState(true);
-  const handleClose = () => {
-    setOpen(false);
-  };
+  // for betaDIalog
+  const { open, handleClose } = useDialog();
 
   React.useEffect(() => {
     window.scrollTo(0, 0);
