+++ client/src/pages/main/Manual.jsx
@@ -1,49 +1,12 @@
-import React, { useState, useEffect } from 'react';
-import axios from '../../utils/axios';
+import React from 'react';
 import AppAppBar from '../../organisms/main/layout/AppAppBar';
 import AppFooter from '../../organisms/main/layout/AppFooter';
 import ProductHero from '../../organisms/main/Main/views/Hero/ProductHero';
 import withRoot from '../../organisms/main/Main/withRoot';
 import textSource from './manualSource/textSource';
 import Manual from '../../organisms/main/Manual/Manual';
-import HOST from '../../utils/config';
-import history from '../../history';
+import useLoginValue from '../../utils/lib/hooks/useLoginValue';
 
-
-const useLoginValue = () => {
-  const [isLogin, setisLogin] = useState(false);
-  const [userType, setUserType] = useState('');
-
-  // logout function
-  const logout = () => {
-    setisLogin(false);
-    axios.get(`${HOST}/api/login/logout`)
-      .then(() => {
-        history.push('/');
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  useEffect(() => {
-    axios.get(`${HOST}/api/login/check`)
-      .then((res) => {
-        if (!res.data.error) {
-          setisLogin(true);
-          setUserType(res.data.userType);
-        } else {
-          setisLogin(false);
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-        setisLogin(false);
-      });
-  });
-
-  return { isLogin, logout, userType };
-};
 
 const MARKETER_TAB_NUMBER = 0;
 const CREATOR_TAB_NUMBER = 1;
